requirements.txt
Flask==3.0.2
pypdf==4.1.0
pyyaml
